Expose Vg's font specifications as records.
And hear Mark Shinwell scream. Given the (low) level at which Vg is
supposed to operate, I don't think we'll ever want fundamentally more
at that level (we could already have only a name and a size). I
promise to come back to Cambridge to pay him a drink at the St
Radeguns if it turns out to be a wrong design decision.